+++ memfile.hpp
@@ -4,13 +4,13 @@
 struct memfile {
 	int fd = 0;
 	char *map = NULL;
-	long long len = 0;
-	long long off = 0;
+	long len = 0;
+	long off = 0;
 	unsigned long tree = 0;
 };
 
 struct memfile *memfile_open(int fd);
 int memfile_close(struct memfile *file);
-int memfile_write(struct memfile *file, void *s, long long len);
+int memfile_write(struct memfile *file, void *s, long len);
 
 #endif
